requirements.txt
# Core
python-telegram-bot==21.7
python-dotenv==1.0.0

# Database
supabase==2.10.0

# AI
google-generativeai==0.8.3

# Security
bcrypt==4.2.0
cryptography==43.0.3

# Google Sheets (optional)
gspread==6.1.4
google-auth==2.36.0

# Utilities
aiohttp==3.11.11
Pillow==11.0.0

# Development
pytest==8.3.4
pytest-asyncio==0.25.0
